+++ src/test/resources/TestNG.xml
@@ -12,7 +12,6 @@
             <class name="Homework18Test"/>
             <class name="Homework19Test"/>
             <class name="Homework21Test"/>
-            <class name="Homework22Test"/>
 
         </classes>
     </test>
